mejoras del programa log

+++ Sistema_Gestion_Documentos.py
@@ -14,12 +14,18 @@ def menu():
 
 extension_archivo={"xls":"Facturas","docx":"Contratos", "pdf":"Informes", "txt":"Backups"}
 
-with open("datos.json", "w") as archivo:
-    json.dump(extension_archivo, archivo)
-
+if not os.path.exists("datos.json"):
+    with open("datos.json", "w") as archivo:
+        json.dump(extension_archivo, archivo)
+   
 with open("datos.json", "r") as archivo:
     datos=json.load(archivo)
 
+def escribir_log(mensaje):
+    with open("log.txt", "a") as archivoLog:
+        now = datetime.now()
+        formatted = now.strftime("[%Y-%m-%d %H:%M:%S]")
+        archivoLog.write(f"{formatted} {mensaje}\n")
 def organizar_archivos():
     for archivo in os.listdir("."):
         if os.path.isfile(archivo) and archivo != "log.txt" and archivo != "datos.json":
@@ -35,10 +41,7 @@ def organizar_archivos():
                             break
                         else:
                             shutil.move(archivo,carpeta)
-                            with open("log.txt", "a") as archivoLog:
-                                now = datetime.now()
-                                formatted = now.strftime("[%Y-%m-%d %H:%M:%S]")
-                                archivoLog.write(f"{formatted} Movido {archivo} a {carpeta}\n")
+                            escribir_log(f"Movido {archivo} a {carpeta}")
                             break
                     except FileExistsError:
                         print("El nombre del archivo ya existe en la carpeta")
@@ -55,10 +58,7 @@ def eliminar_archivo(nombre_archivo):
                     archivoEliminado=os.path.join(ruta_actual,listado)
                     try:
                         os.remove(archivoEliminado)
-                        with open("log.txt", "a") as archivoLog:
-                            now = datetime.now()
-                            formatted = now.strftime("[%Y-%m-%d %H:%M:%S]")
-                            archivoLog.write(f"{formatted} Se ha borrado {nombre_archivo}, se encontraba en: {ruta_actual}\n")
+                        escribir_log(f"Se ha borrado {nombre_archivo}, se encontraba en: {ruta_actual}")
                         break
                     except PermissionError:
                         print("El archivo está en uso o falta privilegios del administrador.")
@@ -69,15 +69,11 @@ def eliminar_archivo(nombre_archivo):
               
 def busqueda_avanzada(nombre_archivo):
     existe=False
-    dic={}
     for ruta_actual, subcarpetas, archivos in os.walk("."):
         for listado in archivos:
             if nombre_archivo.lower() in listado.lower():
                 existe=True
-                dic[listado]=ruta_actual
-    if existe==True:
-        for clave, valor in dic.items():
-            print(f"{clave}, se encuentra en: {valor}")
+                print(f"{listado}, se encuentra en: {ruta_actual}")         
     if existe==False:
         print("No hay un archivo con ese nombre")
 
@@ -102,10 +98,7 @@ def copia_seguridad(archivo):
                     
                     shutil.copy(f"{copiaSeguridad}",f"backup/{nombre}_copia{contador}{extension}")
                     print(f"Se ha hecho una copia de seguridad a {nombre}_copia{contador}{extension}")
-                    with open("log.txt", "a") as archivoLog:
-                        now = datetime.now()
-                        formatted = now.strftime("[%Y-%m-%d %H:%M:%S]")
-                        archivoLog.write(f"{formatted} Se ha creado una copia de seguridad de {nombre}_copia{contador}{extension} a backup\n")
+                    escribir_log(f"Se ha creado una copia de seguridad de {nombre}_copia{contador}{extension} a backup") 
                     break
             if existe==True:
                 break
@@ -120,17 +113,13 @@ def restaurar_copia_seguridad(nombre_archivo):
             nombre, extension =os.path.splitext(nombre_archivo)
             shutil.move(copiaSeguridad,f"./{nombre}_restaurado{extension}")
             print(f"Se ha restaurado el archivo: {nombre}_restaurado{extension}, en el directorio actual")
-            with open("log.txt", "a") as archivoLog:
-                now = datetime.now()
-                formatted = now.strftime("[%Y-%m-%d %H:%M:%S]")
-                archivoLog.write(f"{formatted} Se ha restaurado el archivo {nombre_archivo}_restaurado en el directorio actual.\n")
+
+            escribir_log(f"Se ha restaurado el archivo {nombre_archivo}_restaurado en el directorio actual.") 
+            
         else:
             shutil.move(copiaSeguridad,".")
             print(f"Se ha restaurado el archivo: {nombre_archivo}, en el directorio actual")
-            with open("log.txt", "a") as archivoLog:
-                now = datetime.now()
-                formatted = now.strftime("[%Y-%m-%d %H:%M:%S]")
-                archivoLog.write(f"{formatted} Se ha restaurado el archivo {nombre_archivo} en el directorio actual.\n")
+            escribir_log(f"Se ha restaurado el archivo {nombre_archivo}_restaurado en el directorio actual.")
     else:
         print(f"{nombre_archivo} no está en la carpeta backup")
